requirements.txt
Chameleon==2.11
Jinja2==2.6
Mako==0.7.3
MarkupSafe==0.15
PIL==1.1.7
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
Pygments==1.6rc1
SQLAlchemy==0.8.0b2
Sphinx==1.1.3
WebOb==1.2.3
alembic==0.4.2
amqp==1.0.6
anyjson==0.3.3
anykeystore==0.2
argparse==1.2.1
billiard==2.7.3.19
celery==3.0.13
distribute==0.6.28
docutils==0.10
httplib2==0.7.7
ipython==0.13.1
kombu==2.5.4
oauth2==1.5.211
psycopg2==2.4.6
pyparsing==1.5.6
pyramid==1.4
pyramid-celery==1.3
pyramid-debugtoolbar==1.0.4
pyramid-jinja2==1.6
pyramid-tm==0.7
python-dateutil==1.5
python-openid==2.2.5
redis==2.7.2
repoze.lru==0.6
requests==1.1.0
six==1.2.0
transaction==1.4.0
translationstring==1.1
uWSGI==1.4.5
velruse==1.0.3
venusian==1.0a7
waitress==0.8.2
wsgiref==0.1.2
zope.deprecation==4.0.2
zope.interface==4.0.3
zope.sqlalchemy==0.7.1
